@@ -18,8 +18,8 @@ allprojects {
     }
 
     java {
-		sourceCompatibility = JavaVersion.VERSION_11
-		targetCompatibility = JavaVersion.VERSION_11
+		sourceCompatibility = JavaVersion.VERSION_17
+		targetCompatibility = JavaVersion.VERSION_17
 	}
 
     dependencies {
@@ -28,9 +28,14 @@ allprojects {
             add("api", "org.openapitools:openapi-generator-cli:6.6.0")
         }
 
+        // annotations
+        implementation("org.jetbrains:annotations:16.0.2")
+
+        // testing
         testImplementation("org.junit.jupiter:junit-jupiter-api:5.9.3")
         testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.9.3")
 
+        // logging
         implementation("commons-io:commons-io:2.13.0")
     }
 
