@@ -14,7 +14,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.openapitools.codegen.CodegenConfig;
 import org.openapitools.codegen.CodegenType;
 
-import java.util.Set;
+import java.util.List;
 
 @Slf4j
 public class JavaFeignGenerator extends AbstractCleanJavaCodegen implements CodegenConfig, INitroCodegen {
@@ -54,26 +54,35 @@ public class JavaFeignGenerator extends AbstractCleanJavaCodegen implements Code
         final String apiFolder = JavaCodegenUtils.packageToPath(outputFolder, sourceFolder, apiPackage);
         final String modelFolder = JavaCodegenUtils.packageToPath(outputFolder, sourceFolder, modelPackage);
 
-        // spec package
+        // custom package
         final String specFolder = JavaCodegenUtils.packageToPath(outputFolder, sourceFolder, modelPackage.substring(0, modelPackage.lastIndexOf(".")) + ".spec");
         additionalProperties.put("specPackage", modelPackage.substring(0, modelPackage.lastIndexOf(".")) + ".spec");
+        additionalProperties.put("invokerPackage", modelPackage.substring(0, modelPackage.lastIndexOf(".")));
 
         // ensure directories exist
-        GeneratorUtils.createOutputDirectories(Set.of(invokerFolder, apiFolder, modelFolder, specFolder));
+        GeneratorUtils.createOutputDirectories(List.of(invokerFolder, apiFolder, modelFolder, specFolder));
 
-        // nitro files
-        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("api.peb").targetDirectory(apiFileFolder()).targetFileName("{name}.java").scope(NitroScope.API).iterator(NitroIterator.EACH_API).build());
-        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("api_test.peb").targetDirectory(apiTestFileFolder()).targetFileName("{name}.java").scope(NitroScope.API_TEST).iterator(NitroIterator.EACH_API).overwrite(false).build());
-        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("model.peb").targetDirectory(modelFileFolder()).targetFileName("{name}.java").scope(NitroScope.MODEL).iterator(NitroIterator.EACH_MODEL).build());
-
+        // api files
+        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("api_factory.peb").targetDirectory(invokerFolder).targetFileName("{mainClassName}Factory.java").scope(NitroScope.API).iterator(NitroIterator.ONCE_API).build());
+        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("api_factoryspec.peb").targetDirectory(invokerFolder).targetFileName("{mainClassName}FactorySpec.java").scope(NitroScope.API).iterator(NitroIterator.ONCE_API).build());
         cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("api_main.peb").targetDirectory(apiFileFolder()).targetFileName("{mainClassName}.java").scope(NitroScope.API).iterator(NitroIterator.ONCE_API).build());
         cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("api_collect.peb").targetDirectory(apiFileFolder()).targetFileName("{mainClassName}ApiCollection.java").scope(NitroScope.API).iterator(NitroIterator.ONCE_API).build());
-        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("api_client.peb").targetDirectory(invokerFolder).targetFileName("{mainClassName}Client.java").scope(NitroScope.API).iterator(NitroIterator.ONCE_API).build());
+        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("api_module.peb").targetDirectory(apiFileFolder()).targetFileName("{name}.java").scope(NitroScope.API).iterator(NitroIterator.EACH_API).build());
+        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("api_module_test.peb").targetDirectory(apiTestFileFolder()).targetFileName("{name}.java").scope(NitroScope.API_TEST).iterator(NitroIterator.EACH_API).overwrite(false).build());
 
-        // - spec files
+        // model files
+        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("model.peb").targetDirectory(modelFileFolder()).targetFileName("{name}.java").scope(NitroScope.MODEL).iterator(NitroIterator.EACH_MODEL).build());
         if (cfg.getRequestOverloadSpec()) {
             cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("api_spec.peb").targetDirectory(specFolder).targetFileName("{name}Spec.java").scope(NitroScope.MODEL).iterator(NitroIterator.EACH_API_OPERATION).build());
         }
+
+        // supporting files
+        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("gradle/build.gradle.kts.peb").targetDirectory(outputFolder).targetFileName("build.gradle.kts").scope(NitroScope.SUPPORT).build());
+        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("gradle/settings.gradle.kts.peb").targetDirectory(outputFolder).targetFileName("settings.gradle.kts").scope(NitroScope.SUPPORT).build());
+        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("gradle/gradle-wrapper.properties.peb").targetDirectory(outputFolder).targetFileName("gradle/wrapper/gradle-wrapper.properties").scope(NitroScope.SUPPORT).overwrite(false).build());
+        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("gradle/gradlew.peb").targetDirectory(outputFolder).targetFileName("gradlew").scope(NitroScope.SUPPORT).overwrite(false).build());
+        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("gradle/gradlew.bat.peb").targetDirectory(outputFolder).targetFileName("gradlew.bat").scope(NitroScope.SUPPORT).overwrite(false).build());
+        cfg.addNitroFile(NitroCodegenFile.builder().sourceTemplate("renovate.json.peb").targetDirectory(outputFolder).targetFileName("renovate.json").scope(NitroScope.SUPPORT).build());
     }
 
     /**
