@@ -11,13 +11,12 @@ import java.util.Map;
 public class JavaFeignGeneratorConfig implements INitroCodegenConfig {
 
     public static final String JETBRAINS_ANNOTATION_NULLABLE = "jetbrainsAnnotationsNullable";
-    public static final String REQUEST_OVERLOAD_MAP = "requestOverloadMap";
     public static final String REQUEST_OVERLOAD_SPEC = "requestOverloadSpec";
 
     private Boolean hideGenerationTimestamp = true;
+    private Boolean hideLicense = true;
     private Boolean serializableModel = false;
     private Boolean jetbrainsAnnotationsNullable = true;
-    private Boolean requestOverloadMap = false;
     private Boolean requestOverloadSpec = false;
 
     public static JavaFeignGeneratorConfig of(Map<String, Object> data) {
@@ -25,14 +24,13 @@ public class JavaFeignGeneratorConfig implements INitroCodegenConfig {
 
         // defaults
         cfg.setHideGenerationTimestamp(true);
+        cfg.setHideLicense(true);
         cfg.setSerializableModel(false);
         cfg.setJetbrainsAnnotationsNullable(true);
-        cfg.setRequestOverloadMap(false);
         cfg.setRequestOverloadSpec(true);
 
         // extract from data
         if (data != null) {
-            cfg.setRequestOverloadMap((Boolean) data.get(REQUEST_OVERLOAD_MAP));
             cfg.setRequestOverloadSpec((Boolean) data.get(REQUEST_OVERLOAD_SPEC));
         }
 
